backend/requirements.txt
# backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
bitcoinlib==0.7.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coincurve==21.0.0
cryptography==45.0.5
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
fastecdsa==3.0.1
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
groq==0.30.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==11.3.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynostr==0.6.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.16.0
secp256k1==0.14.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tlv8==0.10.0
tokenizers==0.21.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.9.0
django-redis==5.4.0
pydantic==2.8.2
django-filter==24.2